Projects: hdmi_in: [stable] no changes
Stable Vivado Version: 2015.4

Build Commit, no changes were made

**Notes on this experimental commit type**
I am proposing that we begin using a standard "stable" commit that is
clearly labelled [stable] to indicate that a project is
currently in a state where the hw_handoff and SDK repos have all been
built and tested for the current sources. I feel this is necessary
because the alternative is to require a hardware rebuild and test for
every commit, which would take up a lot of developer time and tempt
people to change multiple things in a single commit. The proposed
policy allows people to make multiple changes to a Vivado project
while only building it once, and yet still follow good versioning
practices by separating the different changes into multiple commits.
The github repo will remain clean and functional as long as we
enforce the rule that whenever a change to a project is pushed the
local repo must be at a commit marked [stable]. Here are some of
the use cases this will benefit:

1) Developer making multiple changes that will require the Vivado
project to be rebuilt:
Developer makes multiples commits to local repo as changes are
made to source. Developer then builds and tests Vivado project.
If test fails, additional changes are made to source and committed
until build/test passes. When test passes, the .hdf and any
changes to the SDK folder are pushed to the local repo in a commit
marked [stable]. The local repo is then pushed to the github repo.

2) Developer making a change that doesn't require the Vivado project to
be rebuilt:
Developer makes change to project that doesn't require a rebuild
(e.g. documentation change, SDK code change). Developer commits
change to local repo with commit marked [stable] that also
describes the changes made. Developer pushes local repo to github
repo.

3) User wanting to use a previous version of a project:
User identifies a previous commit marked [stable] that is built
using the desired version of Vivado. User applies
"git checkout <stable commit> Projects/<project_name>" to checkout
the project at that commit.

4) Developer wants to revert the change made in a previous not stable
commit:
Developer reverts the desired commit to modify the source as
needed. Developer goes through build/test cycle. When test passes,
the .hdf and any changes to the SDK folder are pushed to the local
repo in a commit marked [stable]. The local repo is then pushed to
the github repo.